get new App.jsx to work
